requirements.txt
Flask
flask-wtf
wtforms
requests>=2.0.0
pandas
supabase
python-dotenv
waitress
groq
openai
tiktoken

